+++ oneview_redfish_toolkit/tests/api/test_resource_block_collection.py
@@ -42,6 +42,13 @@ class TestResourceBlockCollection(unittest.TestCase):
         ) as f:
             self.server_hardware_list = json.load(f)
 
+        # Loading ServerProfileTemplate list mockup value
+        with open(
+            'oneview_redfish_toolkit/mockups/oneview/'
+            'ServerProfileTemplates.json'
+        ) as f:
+            self.server_profile_template_list = json.load(f)
+
         # Loading ResourceBlockCollection result mockup
         with open(
             'oneview_redfish_toolkit/mockups/redfish/'
@@ -53,7 +60,7 @@ class TestResourceBlockCollection(unittest.TestCase):
         # Tests if class is correctly instantiated and validated
         try:
             resource_block_collection = ResourceBlockCollection(
-                self.server_hardware_list)
+                self.server_hardware_list, self.server_profile_template_list)
         except Exception as e:
             self.fail("Failed to instantiate ResourceBlockCollection class."
                       " Error: {}".format(e))
@@ -64,7 +71,7 @@ class TestResourceBlockCollection(unittest.TestCase):
         # Tests the serialize function result against known result
         try:
             resource_block_collection = ResourceBlockCollection(
-                self.server_hardware_list)
+                self.server_hardware_list, self.server_profile_template_list)
         except Exception as e:
             self.fail("Failed to instantiate ResourceBlockCollection class."
                       " Error: {}".format(e))
@@ -75,3 +82,19 @@ class TestResourceBlockCollection(unittest.TestCase):
             self.fail("Failed to serialize. Error: {}".format(e))
 
         self.assertEqual(self.resource_block_collection_mockup, result)
+
+    def test_serialize_empty_result(self):
+        expected_result = {
+            "@odata.type": "#ResourceBlockCollection.ResourceBlockCollection",
+            "Name": "Resource Block Collection",
+            "Members@odata.count": 0,
+            "Members": [],
+            "@odata.context": "/redfish/v1/$metadata#ResourceBlockCollection.ResourceBlockCollection",
+            "@odata.id": "/redfish/v1/CompositionService/ResourceBlocks"
+        }
+
+        # Tests the serialize function result against empty list result
+        resource_block_collection = ResourceBlockCollection()
+        result = json.loads(resource_block_collection.serialize())
+
+        self.assertEqual(expected_result, result)