+++ oneview_redfish_toolkit/tests/blueprints/test_resource_block_collection.py
@@ -61,17 +61,6 @@ class TestResourceBlockCollection(unittest.TestCase):
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR,
                 mimetype="application/json")
 
-        @self.app.errorhandler(status.HTTP_404_NOT_FOUND)
-        def not_found(error):
-            """Creates a Not Found Error response"""
-            redfish_error = RedfishError(
-                "GeneralError", error.description)
-            error_str = redfish_error.serialize()
-            return Response(
-                response=error_str,
-                status=status.HTTP_404_NOT_FOUND,
-                mimetype='application/json')
-
         self.app = self.app.test_client()
 
         # propagate the exceptions to the test client
@@ -109,6 +98,11 @@ class TestResourceBlockCollection(unittest.TestCase):
             server_hardware_list = json.load(f)
 
         with open(
+            'oneview_redfish_toolkit/mockups/oneview/ServerProfileTemplates.json'
+        ) as f:
+            server_profile_template_list = json.load(f)
+
+        with open(
             'oneview_redfish_toolkit/mockups/redfish/'
             'ResourceBlockCollection.json'
         ) as f:
@@ -117,29 +111,8 @@ class TestResourceBlockCollection(unittest.TestCase):
         g_mock.oneview_client.server_hardware.get_all.return_value = \
             server_hardware_list
 
-        # Get ResourceBlockCollection
-        response = self.app.get(
-            "/redfish/v1/CompositionService/ResourceBlocks/")
-
-        # Gets json from response
-        expected_result = json.loads(response.data.decode("utf-8"))
-
-        # Tests response
-        self.assertEqual(status.HTTP_200_OK, response.status_code)
-        self.assertEqual("application/json", response.mimetype)
-        self.assertEqual(resource_block_collection_mockup, expected_result)
-
-    @mock.patch.object(resource_block_collection, 'g')
-    def test_get_resource_block_collection_empty(self, g_mock):
-        """Tests ResourceBlockCollection with an empty list"""
-
-        g_mock.oneview_client.server_hardware.get_all.return_value = []
-
-        with open(
-            'oneview_redfish_toolkit/mockups/errors/'
-            'ServerHardwareListNotFound.json'
-        ) as f:
-            server_hardware_list_not_found = json.load(f)
+        g_mock.oneview_client.server_profile_templates.get_all.return_value = \
+            server_profile_template_list
 
         # Get ResourceBlockCollection
         response = self.app.get(
@@ -148,6 +121,34 @@ class TestResourceBlockCollection(unittest.TestCase):
         # Gets json from response
         result = json.loads(response.data.decode("utf-8"))
 
-        self.assertEqual(status.HTTP_404_NOT_FOUND, response.status_code)
+        # Tests response
+        self.assertEqual(status.HTTP_200_OK, response.status_code)
         self.assertEqual("application/json", response.mimetype)
-        self.assertEqual(server_hardware_list_not_found, result)
+        self.assertEqual(resource_block_collection_mockup, result)
+
+    @mock.patch.object(resource_block_collection, 'g')
+    def test_get_resource_block_collection_empty(self, g_mock):
+        """Tests ResourceBlockCollection with an empty list"""
+
+        g_mock.oneview_client.server_hardware.get_all.return_value = []
+        g_mock.oneview_client.server_profile_template.get_all.return_value = []
+
+        # Get ResourceBlockCollection
+        response = self.app.get(
+            "/redfish/v1/CompositionService/ResourceBlocks/")
+
+        # Gets json from response
+        result = json.loads(response.data.decode("utf-8"))
+
+        expected_result = {
+            "@odata.type": "#ResourceBlockCollection.ResourceBlockCollection",
+            "Name": "Resource Block Collection",
+            "Members@odata.count": 0,
+            "Members": [],
+            "@odata.context": "/redfish/v1/$metadata#ResourceBlockCollection.ResourceBlockCollection",
+            "@odata.id": "/redfish/v1/CompositionService/ResourceBlocks"
+        }
+
+        self.assertEqual(status.HTTP_200_OK, response.status_code)
+        self.assertEqual("application/json", response.mimetype)
+        self.assertEqual(expected_result, result)